fix(core): Unified grid spinner into new directive
Various ui-grids where implementing their own spinner elements which all were based on the same
code.  Unified this into a single 'grid-spinner' directive so that we have a central place to manage
the look and placement of the spinner.

The grid-spinner is not limited to grids.  Place it on any element to add a spinner. However, it is
styled to work with the grids, so ymmv.

Added unit test for gridSpinner and fixed myCompanyOrders test

ATLAS-2085