fix(core): Check for allowed state before any state transition
Nwo check to ensure a user is allowed to transition into a state before that trnasition occurs.
Previously, the transition would happen while the acceptance of the Terms of Service was being
tested.  This caused a brief flash of content before the user was redirected to the Unauthorized
state.  It also allowed the user to get into a state where they could not log out from Unauthorized
and get stuck in a loop until they were logged out by the system.

(ATLAS-2086)[https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-2086]