feat(huron): US276958 Customer Admin ability to create a Reseller in Terminus
Currently, if a customer admin tries to add number from customers portal,
the first thing we do is to check if the reseller exists in the Terminus.
If the reseller does not exist we throw an error message to contact the
partner.  The reason was because Terminus previously required the name of
reseller in the Reseller POST V2 API while from Atlas customers portal,
a customer does not have access to the reseller name.
The Terminus API has now been modified so the name is optional. This PR
removes the check and block from UI, so a reseller can be created by a
customer from the customer portal even if the reseller name is absent.