feat(core): US16989 My Subscription: Store DR token whenever Online subscriptions exist
This is the use case where the user is converting a trial into a purchase via the Atlas admin
portal. Currently, DR opens the OPC popup for the user to enter their email address even though
Atlas knows who they are; unfortunately DR does not. Some users are entering the wrong email so the
purchase eventually fails.
This is a short-term solution in which DR would still open the OPC popup
but then OPC would automatically grab the user's SSO token from an Atlas cookie and immediately
redirect back to DR with the token. We'd need to work out the UX of the popup opening and
closing.
The long-term solution is most likely for Atlas to pass the SSO token to BMMP, which passes
to AEM, which passes to DR, meaning no OPC popup.

https://rally1.rallydev.com/#/69751326604d/detail/userstory/120115473280