refactor(hercules): migrate to v2 redirect targets API
The FMS backend is deprecating the v1 redirect target API, and now
provides a new v2 equivalent.  This commit migrates to this new v2
API.

The most notable changes in the backend API are:

 - New endpoints are available for the v2 API;
 - The TTL is not specified by the client anymore;
 - The allowed redirect targets are now called 'allowed registration
   hosts';
 - Each cluster carries its own allowed registration hosts set; ie. it
   is not a global list anymore.

This commit fixes the URLs, removes explicit TTL values, updates the
tests and changes the `setClusterAllowListInfoForExpressway` method in
`hybrid-services-cluster.service.ts` to fetch the allowed registration
hosts for each empty cluster instead of always fetching a global list
of redirect targets and then stitching together the redirect targets
and the clusters.  In the worst case, it may lead to more requests to
the backend (1 per empty Expressway cluster), but in the common case,
it will actually lead to one less HTTP call, since empty Expressway
clusters are not common outside the registration phase.