fix(places): Allow multiple devices per place
With feature toggle 'csdm-multiple-devices-per-place' it is now possible
to add more than one device to a cloudberry place. There are warnings in
the Add Device wizard, in a new modal when clicking the Generate
Activation Code button on a place, and in the device list for a place
with multiple devices unless they are exactly one Spark Voice and one
Spark Share.