WIP(huron): US280721 Delete tab: Implement showing day/time that a file was deleted
Add date/time to deleted page
https://rally1.rallydev.com/#/11303845040ud/detail/userstory/122994081000
In addition, add error code if you attempt to delete a file that is assigned, add a couple of file extensions that will be accepted, and format duration.
All of these were implemented but required small tweaks now that MMS server functionality is available.