feat(core): US16629: Remove temp Change Plan button code for Online release 3.1
Once MC Online 3.1 goes live we want the Change Plan button for MC Online subscriptions to go to the
webex.com Change Order page and for Spark subscriptions to continue to go to the Digital River page.
This PR removes the temporary code that sent MC Online to Digital River while the webex.com page was
being finalized.

https://rally1.rallydev.com/#/69751326604d/detail/userstory/117093604048