fix(huron): Jira 3768 bad API call to terminus
If a customer had not set up an emergency callback number, Atlas was still making a GET call to
terminus, however was passing a string instead of a phone number, which led to a bad API call. Added
a check to verify that customer had an emergency callback number before making API request.