fix(core): DE2555: Plan review page is displayed for customers under a partner
This issue has to do with url parameters stored in sessionStorage that induce a determinant flow in
Atlas Service Setup for pending subscriptions. The solution is to simply clear the storage of the
subscriptionId store in sessionStorage. Another issue related is whenever an Atlas tab is is opened
after crosslaunching into a customer org from an authenticated partner, Atlas directs the user not
to the partner overview but the launched customers overview page. This can be remedied by clearing
the customerOrgId from sessionStorage.

https://rally1.rallydev.com/#/76458539832d/detail/defect/148826739020?fdp=true