fix(devices): Disable possibility to edit multiple services at the same time
Our model does not support easily to edit multiple services for a place at the same time in the edit
services wizard. The solution we chose is to disable this possibility in the ui. So if you edit
calendar, call services is disabled, and if you edit call services, calendar is disabled.