fix(hercules): Preventing the On-Premises Resource Group page from crashing.
Collab-ui's csStatusindicator directive was being used incorrectly in the On-Premises Resource Group page, this was introduced in https://sqbu-github.cisco.com/WebExSquared/wx2-admin-web-client/pull/7002 four days ago. The end result was that the page and all children behaved very strangely, but in a non-deterministic way: Links would just disappear in some cases, depending on the order things were loaded in.

https://jira-eng-chn-sjc1.cisco.com/jira/browse/ATLAS-2497