fix(huron): fix for csSticky not functioning properly
csSticky was sticking to the top of the window as it scrolled and would leave the parent div instead
of stopping at the bottom of the div. Removed sticky polyfill that required javascript on browsers
that did not support css sticky. position: sticky still did not work on any browser, so element was
removed.