fix(trials): DE2048 - (related) fix notes column sorting of expired trials
The notes column is sorted by text with the exception of trials which are lumped together and sorted
by days-remaining.  However, 'expired' and 'expired. purchase now' are two outliers where we no
longer want to sort based on days-remaining.  This commit adds code to the sort algorithm for notes
to identify these two conditions and assign them specific days-remaining values of -2 and -1
respectively.  This results in orderly sorting of the trials items within the larger context of the
entire notes column.