feat(places): Display Hybrid Call numbers on place
If present, show the primaryDn and telephoneNumber of a Hybrid Call
enabled place, under Call Service. If both are present and not the same,
they'll both be under the header for Directory Numbers with a localized
"or" in between them.
Also exposing the SIP URI to make Call Service for Hybrid Call more like
the other two service.