fix(devices): Fix user device settings
Introducing Device API Access caused a problem with settings device
settings like Upgrade Channel and Settings Lockdown on a user. The
reason was that the User object does not have the same id field as a
Place object. Reverting to passing in ids rather than objects, and added
the displayName field that the Device API Access modal needs.

https://jira-eng-gpk2.cisco.com/jira/browse/CE-41580