feat(csdm): Expose more types in suggestion drop down
This adds Network Connectivity, Software Upgrade Channel, Issues and
Tags to the fields that you'll get suggestions for in the drop down. A
lot of changes had to be made to make everything show up localized
properly. I also removed the matching on fields for the suggestions
for now, they can disappear since the search that's done in the
background might not have the letters from the field that you were
typing in them.