feat(csdm): Show fields suggestions for empty search
Search box has auto focus now. Anytime the search box has focus, except
for the first auto focus from entering the page, the drop down is shown.
For an empty current search (there can be existing bullets), suggestions
for all the fields from the aggregations we support are shown. Clicking
on them puts '<field>:' in the search. When '<field>:' is in the current
search, only suggestions for that field are shown.
These feature together make it very easy to navigate with just the mouse!