feat(csdm): Show suggestions based on initial empty search
This means that if you have made a bullet that excludes certain devices,
suggestion that would only hit the devices you excluded will still show
up. This allows you to create "or" bullets, and to make bullets that
result in 0 results until you remove an earlier bullet.
If a suggestion without hits in the current result set is on a field
that you already have a bullet for, we do not show a count. That's
because clicking them will create an "or" bullet, so you might get hits.
If a suggestion without hits in the current result set is not on a field
that you already have a bullet for, we show (0) as the count. That's
because we know that clicking this suggestion will result on 0 results
untill you remove at least one bullet.