fix(csdm): Fix auto focus when returning to browser
I found that if you were autofocused (so no drop down), opened another
app than your browser and then returned to your browser, the code sees
it as another focus event and showed the drop down. I made it so that
only clicking on the search bar or typing while it's focused will show
the drop down on the initial focus.