fix(places): Enable editing call and calendar services
Almost all code was ready for allowing this with correct setting and/or
removal of entitlements, there was just a bug in the merging of state
between wizard steps. If a value in the state was an array, the old and
new value of the array were merged together, making it impossible to
remove a value from the array.

ATLAS-2868