fix(HybridServicesClusterService): Don't stop calculating extended properties when there are no allowed host names in FMS
When building our data structures for hybrid services clusters, we look up the list of redirect targets for empty Expressway clusters, because we display this information in the UI in some cases. FMS's model of redirect targets has changed quite a bit over the years, so in some (rare) cases, we get 404s when asking for redirect targets. This typically happened for very old clusters when registration was never completed. There is no reason to give up in such cases, it is still useful to have all of the other available data for such clusters. In this PR, I just add a .catch() and move on to calculating the other extended cluster properties if we can't get the list of redirect targets.

https://sqbu-github.cisco.com/WebExSquared/fusion-management-service/issues/715 / BEMS730132