chore(*): Upgrade @ciscospark components due to CIAM alert
The spark-js-sdk includes a component (node-jose) that is generating a CIAM vulnerability report,
and the latest version of the sdk still uses a version of the component that is vulnerable.  Fixing
this is a multi-step process.  First, I am upgrading us to the latest version of the SDK to make
sure there are no issues (that is this commit).  Then we will have to make a PR to the sdk component
to get it off the bad version of node-jose, and finally, upgrade Atlas again to use that version.

https://tvce.cisco.com/security/AIMS/AimsAlertPreview.aspx?AlertID=56326