feat(devices): Allow jump to account
Clicking on the icon or "belongs to" name of a device in the right hand
pane will make it jump to the place or users page and open the account
that the device belongs to. Behind 'csdm-device-account-jump' feature
toggle to explore UX possibilities.

Existing 'csdm-device-place-link' feature toggle stays in place, that
toggle enables a breadcrum style navigation from a device. Having both
toggles on is possible.