fix(devices): Remove toggle for device redux
The new search page is GAed for over a week now. Normally we'd wait
longer before cleaning up, but with feature-a suffering from a daily
dose of 503s around US morning time, we don't want customers to run into
the old page anymore. No problems found with the new page.