refactor(core): US27423: injest new API/change license calculation
A new API provides counts of existing licenses by center type. The old way
of calculating existing
licence quantities does not account for transfered
sites, leading to a defect. This checking
injests the new api and refactors the code
to use the data rather than the old calculation, but
keeps the old calculation
in place as a fallback.