fix(devices): Tweak column order
User testing showed that it could be interpreted that the users you saw
in the leftmost column were what the status was about. Moving the device
type to the left so it's more clear what the most important information
in the row is, and adding icon to device column for clarity. Default
sorting now is on the Type column, since that is the first column.

US27522