feat(virtualassistant): CA-2204 - Custom notification when encountering error in ABC CVA setup
Currently, if a user encounters an error while creating a new ABC config, Atlas only shows a generic
"failed to save" notification. This change adds a notification containing a more specific error
message based on the contents of the "error type" field included in the response sent by the Care
Assistant service, if the field is present.

https://jira-eng-sjc12.cisco.com/jira/browse/CA-2204