WIP(csdm): Add search alerts
Add the ability to subscribe to new hits for any search done in the
devices page. The backend currently only sends a single message when
getting a subscribe request, we will implement persistent alerts later.
This is behind a 'csdm-device-search-alerts' feature toggle, UI is
unfinished but this allows us to start testing the functionality.